+++ packages/editor-core/stories/ToolsDrawer/index.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { storyData as mentionStoryData } from '@atlaskit/mention/src/support';
-import { storyData as emojiStoryData } from '@atlaskit/emoji/src/support';
+import { storyData as emojiStoryData } from '@atlaskit/emoji/dist/es5/support';
 import { defaultClientId, defaultServiceHost } from '@atlaskit/media-test-helpers/dist/es5/contextProvider';
 import { defaultCollectionName } from '@atlaskit/media-test-helpers/dist/es5/collectionNames';
 import { StoryBookTokenProvider } from '@atlaskit/media-test-helpers/dist/es5/tokenProvider';