+++ packages/editor-core/stories/renderer/20-table-story.tsx
@@ -18,7 +18,7 @@ import {
 
 import { EmojiProps } from '../../src/renderer/react/nodes/emoji';
 import ProviderFactory from '../../src/providerFactory';
-import { storyData as  emojiStoryData, testData as emojiTestData } from '@atlaskit/emoji/src/support';
+import { storyData as  emojiStoryData, testData as emojiTestData } from '@atlaskit/emoji/dist/es5/support';
 import { toEmojiId } from '@atlaskit/emoji';
 
 import {